Refactor Ast Building
