+++ Code/Interpreter/Tests/AbstractSyntaxTreeTest.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using Antlr4.Runtime;
+﻿using Antlr4.Runtime;
 using Antlr4.Runtime.Tree;
 using Interpreter.Ast;
-using Interpreter.Ast.Nodes.GameObjectNodes;
 using NSubstitute;
 using NUnit.Framework;
 
@@ -12,15 +10,15 @@ namespace Tests
     public class AbstractSyntaxTreeTest
     {
         private IParseTree parseTree;
-        private readonly string parseThisYouFilthyCasual = "./dazel_test_code.txt"; 
+        private const string TestCodePath = "./dazel_test_code.txt";
 
         [SetUp]
         public void Setup()
         {
-            ICharStream stream = CharStreams.fromPath(parseThisYouFilthyCasual);
+            ICharStream stream = CharStreams.fromPath(TestCodePath);
             ITokenSource lexer = new DazelLexer(stream);
             ITokenStream tokens = new CommonTokenStream(lexer);
-            parseTree = (new DazelParser(tokens) {BuildParseTree = true}).start();
+            parseTree = new DazelParser(tokens) {BuildParseTree = true}.start();
         }
         
         
@@ -28,16 +26,11 @@ namespace Tests
         public void AbstractSyntaxTree_Create_VisitGameObject()
         {
             // Arrange
-            IGameObjectVisitor visitor = Substitute.For<IGameObjectVisitor>();
-            
-            AbstractSyntaxTree ast = new(parseTree, visitor);
+            IAstBuilder astBuilder = Substitute.For<IAstBuilder>();
+            astBuilder.BuildAst(parseTree);
 
-            var context = parseTree.GetChild(0) as DazelParser.GameObjectContext;
-            
             // Assert
-            visitor
-                .Received()
-                .VisitGameObject(context);
+            astBuilder.Received().BuildAst(parseTree);
         }
     }
 }
